Update to Handle Scripts Better and CM Servicing Info
Updated CM Scripts section to check for existence of Scripts in the CM environment, including if the feature is enabled or not.
Added details on Update and Servicing for CM.  List update history and what was/was not installed.
Lists all available optional features and if they are turned on or not.
Added variable to capture current path and return to that path at the end of the script.