Links to return to TOC
New function Write-HtmliLink to write a link to an internal location in the document.
New Function Write-HtmlComment to write commented text into the HTML file.  Used to make sure TOC is only in one place.
In custom device collections, list of collections with Maintenance windows, link them to the collection.